Add files via upload
JSP PAGES
JavaScript - It is the one of the most popular web programming language which is used for validations. It is useful in creating the dynamic web page contents and interactive web pages. Along with the server side validation we often use the JavaScript to reduce the load on the server.